Experiments directory
Created experiments directory to hold experimental programs, for testing small portions of the interactive design or trying out different ways to implement a design.

learn is a basic experiment to figure out how to integrate all the core application components and to make sure they work together (Leap API, OpenGL, SDL), on both Linux and Windows.

swipe is an experiment to try out ways to implement a swipe gesture that interacts in real time. It features a left-to-right swiping menu, which may end up being similar to the home screen. So far, the results of this experiment seem to show that the current implementation (strictly physics simulation, translate hand velocity into menu velocity) is too quirky to really be used.