GestureStateGraph tested, nearly finished
The only thing not implemented is removeNode(), which I'm still not sure what the functionality should be - remove any connected edges, or cancel the operation and force the user to remove connections himself?

Tested GestureStateGraph: tested the logical correctness of the implementation, and also designed a few graphs to see how well it worked for actually detecting gestures. Fairly optimistic