3D Carousel prototype
Based on the gesture graph in the "swipe.cpp" test. Need to figure out how to
create a transition from the end of the list to the beginning. Or figure out
a different method for rotation that isn't literally a circle with some finite
elements around it. (Just cycle in the applications that are going to be
displayed at that time?)